refactor: put launch details widgets into widgets folder
