+++ lib/launch_details/widgets/widgets.dart
@@ -1,1 +1,2 @@
-
+export 'details_section.dart';
+export 'image_gallery.dart';
